% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello!}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{character string to be added to "Hello, " string.}
}
\description{
A function that says hello.
}
\examples{
hello("world")
}

