% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlalomModel-methods.R
\name{init}
\alias{init}
\title{#' Initialize a SlalomModel object
#'
#' Initialize a SlalomModel with sensible starting values for parameters before
#' training the model.
#'
#' @param object a \code{Rcpp_SlalomModel} object
#' @param alpha_priors numeric(2) giving alpha and beta hyperparameters for a
#' gamma prior distribution for alpha parameters (precision of factor weights)
#' @param epsilon_priors numeric(2) giving alpha and beta hyperparameters for a
#' gamma prior distribution for noise precision parameters
#' @param noise_model character(1) defining noise model, defaults to "gauss" for
#' Gaussian noise model
#' @param seed integer(1) value supplying a random seed to make results
#' reproducible
#' @param ... generic arguments passed to \code{Rcpp_SlalomModel} method
#'
#' @details It is strongly recommended to use \code{\link{newSlalomModel}} to
#' create the \code{\link{SlalomModel}} object prior to applying
#' \code{initialize}.
#'
#' @docType methods
#' @name init
#' @rdname init
#' @aliases init init,Rcpp_SlalomModel-method
#'
#' @author Davis McCarthy
#' @importFrom rsvd rpca
#' @importFrom stats runif
#' @importFrom stats rnorm
#' @export
#' @examples
#' gmtfile <- system.file("extdata", "reactome_subset.gmt", package = "slalom")
#' genesets <- GSEABase::getGmt(gmtfile)
#' data("mesc")
#' model <- newSlalomModel(mesc, genesets, n_hidden = 5, min_genes = 10)
#' model <- init(model)
## setGeneric("init", function(object, ...) standardGeneric("init"))}
\usage{
init(object, alpha_priors = NULL, epsilon_priors = NULL,
  noise_model = "gauss", seed = NULL, pi_prior = NULL, n_hidden = NULL,
  design = NULL, drop_factors = TRUE)
}
\arguments{
\item{object}{a \code{Rcpp_SlalomModel} object}

\item{alpha_priors}{numeric(2) giving alpha and beta hyperparameters for a
gamma prior distribution for alpha parameters (precision of factor weights)}

\item{epsilon_priors}{numeric(2) giving alpha and beta hyperparameters for a
gamma prior distribution for noise precision parameters}

\item{noise_model}{character(1) defining noise model, defaults to "gauss" for
Gaussian noise model}

\item{seed}{integer(1) value supplying a random seed to make results
reproducible}

\item{pi_prior}{numeric matrix (genes x factors) giving prior probability of
a gene being active for a factor}

\item{n_hidden}{integer(1), number of hidden factors in model. Required if
\code{pi_prior} is not \code{NULL}.}

\item{design}{design matrix of covariates/factors to fit as known factors
in the model.}

\item{drop_factors}{logical(1), should factors be dropped from the model if
the model determines them not to be relevant? Default is \code{TRUE}.}
}
\value{
an `Rcpp_SlalomModel` object
}
\description{
#' @name init
#' @rdname init
#' @docType methods
#' #@export
## setMethod("init", "Rcpp_SlalomModel", function(
Initialize a SlalomModel object
}
\details{
Initialize a SlalomModel with sensible starting values for parameters before
training the model.

It is strongly recommended to use \code{\link{newSlalomModel}} to
create the \code{\link{SlalomModel}} object prior to applying
\code{initialize}.
}
\examples{
gmtfile <- system.file("extdata", "reactome_subset.gmt", package = "slalom")
genesets <- GSEABase::getGmt(gmtfile)
data("mesc")
model <- newSlalomModel(mesc, genesets, n_hidden = 5, min_genes = 10)
model <- init(model)
}
\author{
Davis McCarthy
}
