% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello}
\alias{rcpp_hello}
\title{rcpp_hello}
\usage{
rcpp_hello()
}
\value{
a list with a character vector and a numeric vector
}
\description{
This is a simple function using Rcpp that creates an R list containing a
character vector and a numeric vector.
}
\details{
Not much to report.
}
\examples{
rcpp_hello()
}

