% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Rcpp_SlalomModel}
\alias{Rcpp_SlalomModel}
\alias{Rcpp_SlalomModel-class}
\title{The "Slalom Model" (SlalomModel)  class}
\description{
S4 class and the main class used by slalom to hold model data and results.
SingleCellExperiment extends the Bioconductor SummarizedExperiment class.
}
\details{
This class is initialized from a matrix of expression values and a collection
of genesets in a \code{GeneSetCollection} object from the GSEABase package.

Methods that operate on SingleCellExperiment objects constitute the basic scater workflow.
}
\section{Methods}{

\describe{
\item{\code{train()}}{void train()
docstring : Train the SlalomModel}

\item{\code{update()}}{void update()
docstring : Update the SlalomModel}

\item{\code{updateAlpha(...)}}{void updateAlpha(int)
docstring : Update alpha}

\item{\code{updateEpsilon()}}{void updateEpsilon()
docstring : Update Epsilon}

\item{\code{updatePi(...)}}{void updatePi(int)
docstring : Update Pi}

\item{\code{updateW(...)}}{void updateW(int)
docstring : Update W}

\item{\code{updateX(...)}}{void updateX(int)
docstring : Update X}
}}

\section{Slots}{

 \describe{
   \item{\code{.xData}:}{Environment enabling access to the C++-level
   SlalomModel object.}
}
}

