% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slalom-package.R
\docType{package}
\name{slalom}
\alias{slalom}
\alias{slalom-package}
\title{Factorial single-cell latent variable models}
\description{
slalom
}
\details{
Factorial latent variable models for RNA-seq data.
}
\author{
Davis McCarthy
}
