% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SlalomModel}
\alias{SlalomModel}
\title{SlalomModel C++ class}
\arguments{
\item{Y_init}{matrix of expression values}

\item{pi_init}{G x K matrix with each entry being the prior
probability for a gene g being active for factor k.}

\item{X_init}{matrix of initial factor states (N x K)}

\item{W_init}{G x K matrix of initial weights}

\item{prior_alpha}{numeric vector of length two giving prior values
for the gamma hyperparameters of the precisions}

\item{prior_epsilon}{numeric vector of length two giving prior values
for the gamma hyperparameters of the residual variances}
}
\value{
an object of the SlalomModel class
}
\description{
A C++ class for SlalomModel models.
}
