% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlalomModel-methods.R
\docType{methods}
\name{train}
\alias{train}
\alias{train,Rcpp_SlalomModel-method}
\alias{train}
\title{Train a SlalomModel object}
\usage{
train(object, ...)

\S4method{train}{Rcpp_SlalomModel}(object, nIterations = 5000,
  tolerance = 1e-08, forceIterations = FALSE, minIterations = 700,
  shuffle = TRUE)
}
\arguments{
\item{object}{a \code{Rcpp_SlalomModel} object}

\item{...}{generic arguments passed to \code{Rcpp_SlalomModel} method}

\item{nIterations}{integer(1) maximum number of iterations to use in training
the model (default: 5000)}

\item{tolerance}{numeric(1) tolerance to allow between iterations
(default 1e-08)}

\item{forceIterations}{logical(1) should the model be forced to update
\code{nIteration} times?}

\item{minIterations}{integer(1) minimum number of iterations to perform.}

\item{shuffle}{logical(1) should the order in which factors are updated be
shuffled between iterations? Shuffling generally helps speed up convergence
so is recommended and defaults is \code{TRUE}}
}
\value{
an `Rcpp_SlalomModel` object
}
\description{
Train a SlalomModel to infer model parameters.
}
\details{
Train the model using variational Bayes methods to infer parameters.
}
\examples{
gmtfile <- system.file("extdata", "reactome_subset.gmt", package = "slalom")
genesets <- GSEABase::getGmt(gmtfile)
data("mesc")
model <- newSlalomModel(mesc, genesets, n_hidden = 5, min_genes = 10)
model <- init(model)
model <- train(model, nIterations = 10)
}
\author{
Davis McCarthy
}
