% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlalomModel-methods.R
\docType{methods}
\name{init}
\alias{init}
\alias{init,SlalomModel-method}
\title{Initialize a SlalomModel object}
\usage{
init(object, ...)

\S4method{init}{SlalomModel}(object, alpha_priors = NULL,
  epsilon_priors = NULL, noise = "gauss")
}
\arguments{
\item{object}{a \code{SlalomModel} object}

\item{...}{generic arguments passed to}

\item{alpha_priors}{numeric(2) giving alpha and beta hyperparameters for a
gamma prior distribution for alpha parameters (precision of factor weights)}

\item{epsilon_priors}{numeric(2) giving alpha and beta hyperparameters for a
gamma prior distribution for noise precision parameters}

\item{noise}{character(1) defining noise model, defaults to "gauss" for
Gaussian noise model}
}
\description{
Initialize a SlalomModel with sensible starting values for parameters before
training the model.
}
\details{
It is strongly recommended to use \code{\link{newSlalomModel}} to
create the \code{\link{SlalomModel}} object prior to applying
\code{initialize}.
}
\examples{
gmtfile <- system.file("extdata", "reactome_subset.gmt", package = "slalom")
genesets <- GSEABase::getGmt(gmtfile)
data("mesc")
model <- newSlalomModel(mesc, genesets, n_hidden = 5, min_genes = 10)
model <- init(model)
}
\author{
Davis McCarthy
}

