% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlalomModel-methods.R
\name{newSlalomModel}
\alias{newSlalomModel}
\title{Create a new SlalomModel object.}
\usage{
newSlalomModel(object, genesets, n_hidden = 5, prune_genes = TRUE,
  min_genes = 15)
}
\arguments{
\item{object}{\code{"SCESet"} object N x G expression data matrix (cells x genes)}

\item{genesets}{a \code{"GeneSetCollection"} object containing annotated
gene sets}

\item{n_hidden}{number of hidden factors to fit in the model (2-5 recommended)}

\item{prune_genes}{logical, should genes that are not annotated to any gene
sets be filtered out?}

\item{min_genes}{scalar, minimum number of genes required in order to retain
a gene set for analysis}
}
\value{
a new SlalomModel object
}
\description{
Slalom fits relatively complicated hierarchical Bayesian factor analysis
models with data and results stored in a \code{"SlalomModel"} object. This
function builds a new \code{"SlalomModel"} object from minimal inputs.
}
\details{
This function builds and returns the object, checking for validity, which
includes checking that the input data is of consistent dimensions.
}
\examples{
gmtfile <- system.file("extdata", "reactome_subset.gmt", package = "slalom")
genesets <- GSEABase::getGmt(gmtfile)
data("mesc")
model <- newSlalomModel(mesc, genesets, n_hidden = 5, min_genes = 10)
}

